continued styling - removed schemaWindow and rolled into Container
