+++ client/components/URIInput.jsx
@@ -2,21 +2,21 @@ import React from 'react';
 import { TextField, Button } from '@mui/material';
 import InputAdornment from '@mui/material/InputAdornment';
 import Computer from '@mui/icons-material/Computer';
+import '../Styles/Dashboard.css';
 
 function URIInput(props) {
 
     const { uri, setUri } = props;
 
     return (
-        <div className='URIInput'>
+        <div className='uri-input'>
             <form>
                 <label form='postgreSQL'>PostgreSQL URI:</label><br></br>
                 <TextField
                     variant='outlined'
                     size='small'
                     type='text'
-                    id='URIAddr'
-                    name='URIAddr'
+                    className='uri-text'
                     InputProps={{
                         startAdornment: (
                             <InputAdornment position="start">
@@ -29,7 +29,7 @@ function URIInput(props) {
                     type='button'
                     size='large'
                     variant='contained'
-                    id='uri-button'>
+                    className='uri-button'>
                     Go!
                 </Button>
             </form>
